made the tests part of integration tests
