+++ build.gradle
@@ -53,10 +53,19 @@ sourceSets {
             srcDir('src/IntegrationTests/java')
         }
     }
+    e2e {
+        java {
+            compileClasspath += main.output + test.output
+            runtimeClasspath += main.output + test.output
+            srcDir('src/e2e/java')
+        }
+    }
+
 }
 
 dependencies {
     integrationTestCompile(sourceSets.test.output)
+    e2eCompile(sourceSets.test.output)
 }
 
 task functional(type: Test) {
@@ -117,6 +126,8 @@ sonarqube {
 configurations {
     integrationTestCompile.extendsFrom(testCompile)
     integrationTestRuntime.extendsFrom(testRuntime)
+    e2eRuntime.extendsFrom(testRuntime)
+    e2eCompile.extendsFrom(testCompile)
 }
 
 bootRun {
@@ -187,6 +198,11 @@ run {
     }
 }
 
-task smoke(type: Test, description: 'Runs the smoke tests.', group: 'Verification') {
-
+task smoke(type: Test) {
+    group = 'verification'
+    description = 'Executes non-destructive smoke tests against a running case loader'
+    setTestClassesDirs(sourceSets.e2e.output.classesDirs)
+    setClasspath(sourceSets.e2e.runtimeClasspath)
+    include "uk/gov/hmcts/reform/sscs/smoke/**"
+    exclude "uk/gov/hmcts/reform/sscs/functional/**"
 }
