+++ src/main/resources/config/application.properties
@@ -15,7 +15,7 @@ claiming.expenses.link=${CLAIMING_EXPENSES_LINK:http://localhost:3000/expenses/a
 gov.uk.notification.api.key=${NOTIFICATION_API_KEY:emailnotificationkey}
 subscription.hmac.secret.text=${EMAIL_MAC_SECRET_TEXT:our-big-secret}
 
-idam.s2s-auth.url=${S2S_URL:}
+idam.s2s-auth.url=${S2S_URL:http://localhost:4502}
 
 job.scheduler.enabled=${JOB_SCHEDULER_ENABLED:false}
 job.scheduler.url=${JOB_SCHEDULER_URL:http://localhost:8484}
@@ -51,7 +51,6 @@ notification.evidenceReminder.smsId=${EVIDENCE_REMINDER_SMS_TEMPLATE_ID:e3f71440
 idam.url=${IDAM_URL:http://localhost:4501}
 idam.s2s-auth.totp_secret=${IDAM_S2S-AUTH_TOTP_SECRET:AAAAAAAAAAAAAAAC}
 idam.s2s-auth.microservice=${IDAM_S2S-AUTH_MICROSERVICE:sscs}
-idam.s2s-auth.url=${IDAM_S2S-AUTH:http://localhost:4502}
 idam.oauth2.user.email=${IDAM_SSCS_SYSTEMUPDATE_USER:SSCS_SYSTEM_UPDATE}
 idam.oauth2.user.password: ${IDAM_SSCS_SYSTEMUPDATE_PASSWORD:SSCS_SYSTEM_UPDATE}
 idam.oauth2.client.id=${IDAM_OAUTH2_CLIENT_ID:sscs}
